Main screen turn on
The backlight was not staying on because the PWM pulse cycle
was setting the incorrect register (it was setting the register
in the pwm region instead of rpwm) causing the cycle to be 0
and the lcd appear to not be initialized.

It now turns on with a white screen, indicating a problem with
the display engine initialization, but the backlight and lcd are
being initialized properly.

Start adding code to debug.