Work in progress - initialize display driver correctly?
This updates the display driver to fix a few bugs where it
was accessing the wrong address space for MIXER0 in the DE.

It *seems* to work better, because after AHB_RESET the display
turns white (as before) but after it enables the framebuffer turns
black. It may be that the FB is allocated with malloc and it needs
to be in uncached memory, but ucmalloc/fbmemalloc panics (there is
no memory set aside as adventuresin9 said on 9fans.)

The register writes have been compared to the NuttX/p-boot implementation
documented in https://lupyuen.github.io/articles/de#appendix-initialising-the-allwinner-a64-display-engine and seem to match.