Fix /dev/keyadc_event
Use a Rendez point instead of spinloop for waking up from interrupts.
This is both more efficient and more robust. Before, it would only return
the first interrupt and then spin forever.

Also fixes a bug where /dev/arch was returning the wrong data after the first
read because the offset for readstr was wrong.